+++ NEWS.md
@@ -1,3 +1,7 @@
-# PACKAGE 2020.2.17
+# splutil 2022.4.28
 
-- Package is created
+- split_equal (Split a vector into a list with equal sized elements)
+- is_fully_named_list (Is this a fully named list?)
+- all_list_elements_null_or_df (Are all elements in a list null or data.frames?)
+- all_list_elements_null_or_list (Are all elements in a list null or lists?)
+- all_list_elements_null_or_fully_named_list (Are all elements in a list null or fully named lists?)